refactor(worldgen): RiverWidth newtype + fix all clippy -D warnings
RiverWidth(pub f32) wraps river-width values with a clamping constructor,
making the [MIN_RIVER_WIDTH, MAX_RIVER_WIDTH] invariant visible at write
sites (FineRegion.width, RiverSegment.width). Arithmetic uses explicit
.0 deref to be honest about unwrapping. From<f32> delegates to new().

Fixes all cargo clippy --all-targets -- -D warnings across the codebase:
- worldgen: collapsible-if → let-chains, redundant casts, map_or →
  is_some_and/is_none_or, needless range loops, closure → function,
  too_many_arguments allow-attrs on geological pipeline fns
- mesher, voxel: .contains() instead of iter().any()
- render, jobs, ecs: too_many_arguments, dead_code, let-chains
- ui: collapsible-if, is_multiple_of
- persistence: items_after_test_module (moved Drop before test block)
- worldgen_viz: egui from_id_source → from_id_salt, collapsible ifs,
  derive(Default), dead_code

328 tests pass. Fingerprint 0x0652_5CEE_A30F_9710 unchanged.

+++ src/worldgen/caves/system.rs
@@ -131,6 +131,7 @@ fn roll_bounding_box(
 /// already-placed chamber. Sampling stops after `max_tries = 200` attempts
 /// regardless of whether the target count was reached — the system uses
 /// however many fit, down to a minimum of 1.
+#[allow(clippy::too_many_arguments)]
 fn sample_chambers(
     seed: u64,
     coord: RegionCoord,
@@ -363,6 +364,7 @@ fn connect_chambers_mst(
 /// if the chamber top is `SKYLIGHT_DEPTH_MIN–SKYLIGHT_DEPTH_MAX` below
 /// the surface. Exactly one entrance per chamber; priority Sinkhole →
 /// CliffMouth → Skylight.
+#[allow(clippy::too_many_arguments)]
 fn roll_entrances(
     seed: u64,
     coord: RegionCoord,
@@ -393,7 +395,7 @@ fn roll_entrances(
             entrances.push(Entrance {
                 chamber_idx: ci as u32,
                 kind: EntranceKind::Sinkhole,
-                surface: IVec3::new(cwx, surface_h + SURFACE_BAND as i32, cwz),
+                surface: IVec3::new(cwx, surface_h + SURFACE_BAND, cwz),
             });
             continue;
         }
@@ -427,7 +429,7 @@ fn roll_entrances(
             entrances.push(Entrance {
                 chamber_idx: ci as u32,
                 kind: EntranceKind::Skylight,
-                surface: IVec3::new(cwx, surface_h + SURFACE_BAND as i32, cwz),
+                surface: IVec3::new(cwx, surface_h + SURFACE_BAND, cwz),
             });
         }
     }
